fix: rpt_stock fields to double

+++ lib/features/sales/domain/entities/check_stock.dart
@@ -6,10 +6,10 @@ class CheckStockEntity {
   final String itemName;
   final String storeCode;
   final String storeName;
-  final int qtyOnHand;
-  final int ordered;
-  final int commited;
-  final int available;
+  final double qtyOnHand;
+  final double ordered;
+  final double commited;
+  final double available;
 
   CheckStockEntity({
     required this.itemCode,
@@ -27,10 +27,10 @@ class CheckStockEntity {
     String? itemName,
     String? storeCode,
     String? storeName,
-    int? qtyOnHand,
-    int? ordered,
-    int? commited,
-    int? available,
+    double? qtyOnHand,
+    double? ordered,
+    double? commited,
+    double? available,
   }) {
     return CheckStockEntity(
       itemCode: itemCode ?? this.itemCode,
@@ -63,10 +63,10 @@ class CheckStockEntity {
       itemName: map['itemName'] as String,
       storeCode: map['storeCode'] as String,
       storeName: map['storeName'] as String,
-      qtyOnHand: map['qtyOnHand'] as int,
-      ordered: map['ordered'] as int,
-      commited: map['commited'] as int,
-      available: map['available'] as int,
+      qtyOnHand: map['qtyOnHand'] as double,
+      ordered: map['ordered'] as double,
+      commited: map['commited'] as double,
+      available: map['available'] as double,
     );
   }
